:recycle: refactor move CertificationCenterInvitedUser to team
