+++ api/src/team/domain/models/OrganizationInvitedUser.js
@@ -1,10 +1,7 @@
-import {
-  AlreadyAcceptedOrCancelledInvitationError,
-  AlreadyExistingMembershipError,
-  NotFoundError,
-} from '../../../shared/domain/errors.js';
+import { AlreadyExistingMembershipError, NotFoundError } from '../../../shared/domain/errors.js';
 import { roles } from '../../../shared/domain/models/Membership.js';
 import { OrganizationInvitation } from '../../../team/domain/models/OrganizationInvitation.js';
+import { AlreadyAcceptedOrCancelledInvitationError } from '../errors.js';
 
 class OrganizationInvitedUser {
   constructor({ userId, invitation, currentRole, organizationHasMemberships, currentMembershipId, status } = {}) {
