+++ api/src/team/infrastructure/repositories/certification-center-invited-user.repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../shared/domain/errors.js';
-import { CertificationCenterInvitedUser } from '../../../shared/domain/models/CertificationCenterInvitedUser.js';
+import { CertificationCenterInvitedUser } from '../../domain/models/CertificationCenterInvitedUser.js';
 
 const get = async function ({ certificationCenterInvitationId, email }) {
   const invitation = await knex('certification-center-invitations')
