+++ api/tests/integration/application/error-manager_test.js
@@ -7,6 +7,7 @@ import {
 } from '../../../src/identity-access-management/domain/errors.js';
 import { CampaignParticipationDeletedError } from '../../../src/prescription/campaign-participation/domain/errors.js';
 import * as DomainErrors from '../../../src/shared/domain/errors.js';
+import { AlreadyAcceptedOrCancelledInvitationError } from '../../../src/team/domain/errors.js';
 import { expect, HttpTestServer, sinon } from '../../test-helper.js';
 
 describe('Integration | API | Controller Error', function () {
@@ -259,7 +260,7 @@ describe('Integration | API | Controller Error', function () {
     });
 
     it('responds Conflict when an AlreadyAcceptedOrCancelledInvitationError occurs', async function () {
-      routeHandler.throws(new DomainErrors.AlreadyAcceptedOrCancelledInvitationError());
+      routeHandler.throws(new AlreadyAcceptedOrCancelledInvitationError());
       const response = await server.requestObject(request);
 
       expect(response.statusCode).to.equal(CONFLICT_ERROR);
