+++ api/tests/team/integration/infrastructure/repositories/certification-center-invited-user.repository.test.js
@@ -1,8 +1,8 @@
 import lodash from 'lodash';
 
 import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
-import { CertificationCenterInvitedUser } from '../../../../../src/shared/domain/models/CertificationCenterInvitedUser.js';
 import { CertificationCenterInvitation } from '../../../../../src/team/domain/models/CertificationCenterInvitation.js';
+import { CertificationCenterInvitedUser } from '../../../../../src/team/domain/models/CertificationCenterInvitedUser.js';
 import { certificationCenterInvitedUserRepository } from '../../../../../src/team/infrastructure/repositories/certification-center-invited-user.repository.js';
 import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
